# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Rental Car POS & Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [unpaid, paid, partial, null]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction / invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction (e.g. mark as paid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ─── REPORTS ──────────────────────────────────────────────────────────────────
...
  /reports/revenue:
    get:
      operationId: getRevenueReport
      tags: [reports]
      summary: Revenue report over a date range
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: ["string", "null"]
            enum: [day, week, month, null]
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDataPoint"
  /reports/vehicles:
    get:
      operationId: getVehicleUtilizationReport
      tags: [reports]
      summary: Vehicle utilization report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vehicle utilization
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VehicleUtilizationReport"
  /reports/partners:
    get:
      operationId: getPartnerReport
      tags: [reports]
      summary: Partner (mitra) profit sharing report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Partner report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartnerReport"
  /reports/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [reports]
      summary: Recent booking and transaction activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ─── USERS ────────────────────────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List system users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        invoiceNumber:
          type: string
        customerName:
          type: string
        vehicleName:
          type: string
        amount:
          type: number
        paidAmount:
          type: number
        status:
          type: string
          enum: [unpaid, paid, partial]
        paymentMethod:
          type: ["string", "null"]
          enum: [cash, transfer, qris, null]
        paidAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - bookingId
        - invoiceNumber
        - customerName
        - vehicleName
        - amount
        - paidAmount
        - status
        - paymentMethod
        - paidAt
        - notes
        - createdAt
    CreateTransactionBody:
      type: object
      properties:
        bookingId:
          type: integer
        amount:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - bookingId
        - amount
    UpdateTransactionBody:
      type: object
      properties:
        paidAmount:
          type: number
        status:
          type: string
          enum: [unpaid, paid, partial]
        paymentMethod:
          type: ["string", "null"]
          enum: [cash, transfer, qris, null]
        notes:
          type: ["string", "null"]
...
    RevenueDataPoint:
      type: object
      properties:
        period:
          type: string
        revenue:
          type: number
        bookingCount:
          type: integer
      required:
        - period
        - revenue
        - bookingCount
    VehicleUtilizationReport:
      type: object
      properties:
        vehicleId:
          type: integer
        vehicleName:
          type: string
        plateNumber:
          type: string
        ownership:
          type: string
        totalBookings:
          type: integer
        totalRevenue:
          type: number
        utilizationDays:
          type: integer
        utilizationPercent:
          type: number
      required:
        - vehicleId
        - vehicleName
        - plateNumber
        - ownership
        - totalBookings
        - totalRevenue
        - utilizationDays
        - utilizationPercent
    PartnerReport:
      type: object
      properties:
        vehicleId:
          type: integer
        vehicleName:
          type: string
        plateNumber:
          type: string
        partnerName:
          type: string
        totalRevenue:
          type: number
        operationalCost:
          type: number
        netRevenue:
          type: number
        partnerShare:
          type: number
        companyShare:
          type: number
        profitSharePercent:
          type: number
      required:
        - vehicleId
        - vehicleName
        - plateNumber
        - partnerName
        - totalRevenue
        - operationalCost
        - netRevenue
        - partnerShare
        - companyShare
        - profitSharePercent
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [booking_created, checkin, checkout, payment, maintenance]
        description:
          type: string
        relatedId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - type
        - description
        - relatedId
        - createdAt
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, staff, owner]
        createdAt:
          type: string
      required:
        - id
        - username
        - name
        - role
        - createdAt
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, staff, owner]
      required:
        - username
        - name
        - password
        - role
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user